requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
pandas==2.1.4
python-binance==1.0.19
google-generativeai==0.3.2 